backend/requirements.txt
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
cloudevents==1.8.0
colorama==0.4.6
deprecation==2.1.0
et-xmlfile==1.1.0
Flask==2.2.2
functions-framework==3.3.0
google-api-core==2.11.0
google-auth==2.15.0
google-cloud==0.34.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.3
googleapis-common-protos==1.57.0
grpcio==1.51.1
grpcio-status==1.51.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.24.1
openai==0.25.0
openpyxl==3.0.10
packaging==22.0
pandas==1.5.2
pandas-stubs==1.5.2.221213
proto-plus==1.22.1
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
rsa==4.9
six==1.16.0
tqdm==4.64.1
types-pytz==2022.7.0.0
typing_extensions==4.4.0
urllib3==1.26.13
watchdog==2.2.0
Werkzeug==2.2.2
